Reworked the look and drop commands and added a take command
Now the game no longer asks you which items you want to drop or pick up, you just specify that in the command itself and you can specify as many items as you want. Look command now only displays the long room description and the items in the room. The command class instead of having a secondword parameter now has a arguments array