Code refactoring
Expanded the base Biome class to have a name and the GameManager class to have a property of type biome to be able to reference current biome methods easier, such as checking for quests, which now can just be done by using the function instead of checking for the specific biome and invoking its specifi function. Also implemented showing the map for the mountains biome as well as updated rooms. I updated the remaining biomes with the BiomeName variable, so that the biome welcome message functions properly